created the starting skeleton for the athlete profile
