+++ client/.babelrc
@@ -1,0 +1,10 @@
+{
+  "presets": [
+    "env",
+    "stage-2",
+    "react"
+  ],
+  "plugins": [
+    "syntax-async-functions"
+  ]
+}
